fix(audio): display autoplay modal only after initial setup is done
If the audio modal is open, the user hasn't gone through the audio setup
flow yet, which generally "fixes" the autoplay issue by itself (generally
due to gUM being requested). We currently do not wait for that to happen,
though, which means an unecessary prompt might be shown. Additionally,
the current prompt strategy may interrupt the audio configuration
process, even if momentarily, whic is not ideal.

Wait for the audio modal to be closed before showing the autoplay prompt
component. The storage key override in audio-manager's onAudioJoin callback
is also removed for two reasons:
  - LiveKit join procedure may happen concurrently with the config
    process, and associating modal closure with audio joining is
    incorrect
  - The original reason for that override seems to not be valid anymore.