fix(webcam): clean up local connecting streams when they're shared
The local "connectingStream" state is not getting cleaned up when the
camera is effectively shared (i.e.: it's present in the remote streams
list). I do not know whether this was intentional or not, but it does
more harm (inconsistent client-server state when a server side ejection
happens) than good (a *very* small chance that this keeps the camera up
during a reconnection). As of now, I still deem state consistency more
important due to how camera states are handled all around. Better
cam re-connection should be dealt with holistically later.

Clean up the local connectingStream state whenever the camera is
effectively shared. This fixes a couple of scenarios that could cause a
client-server state inconsitency - the user thinks their camera is
shared, but it's not.