fix(livekit): unable to unmute when sharing screen+audio beforehand
If an user shares a Chrome tab with audio before unmuting their mic for
the first time, unmuting will be a no-op. This is caused by the LiveKit
audio bridge not filtering publications for ther proper source types in
all scenarios.

Refactor local track publication in LiveKit's audio bridge so that it's
centralized and guarantee that tracks are filtered for mic sources only.