fix(livekit): do not reset inputStream if its device didn't change
LiveKit's audio bridge causes a re-publish whenever its inputStream
change. There's a check in there to prevent unnecessary republishes, but
since it only compares the media stream ID, it's always going to be different
as the internal LiveKit media stream is a cloned one. This causes a few
client bugs as: mute toggling, duplicate published streams, etc.

Enrich the input stream reset check to also look into whether the
stream's deviceIds changed. The only exception to this rule is when
audio constraints are updated, so an optional parameter to force the
reset was added as well.