fix(livekit): cleanup original input stream when "exiting" audio
LiveKit's bridge implements the exitAudio interface, even though that
method is not fully functional in the bridge. It's in use, however, and
the fact that it won't cleanup the original stream when audio "exits"
may cause the microphone to be inoperable due to new being stuck with a
stale media stream.

Guarantee that the exitAudio method in LiveKit's audio bridge properly
cleanup the local originalStream prop.