fix(screenshare): subscription occasionally fails
LiveKit's screen share bridge occasionally fails to subscribe to remote
publications - more likely when joining a room with an ongoing
subscription. There was a routine in place to sync publications on join,
but it had an issue where publication _tracks_ might no be populated yet
when running that sync. That causes the source check for the track to
fail, and so does the subscription.

Improve the chances screen subscription works by:
  - Checking the publication for its source type when doing the initial
    sync rather than the track. The publication is more likely to be
    populated at that point.
  - Add a delayed subscription retry (1s) whenever the initial subscribe
    fails due to a "Publication not found" error.