fix(layout): add backdrop to the side panels
Adds a backdrop to the sidebar navigation and sidebar panel so that they
hide the things behind. This started as a fix for the warning banner but
turned out to be a bigger thing as it also adds a minimum margin between
the sidebar panel and the media area. Additionaly it enhances the
sidebar navigation for mobile by putting a little shadow box so that the
floating effect is more realistic and also fixes the toggle button to
hide only the navigation sidebar.