fix(crowd-action-button): change style, size and locale
Changes the style of the label of the buttons in the bottom of the user
list panel to match the correct size and color.
Adds maximum height to those buttons, text overflow to its labels and
tooltip with more detailed description in case the label is not entirely
shown.
Changes mute all except presenter locale to something shorter.