fix(audio-captions): selector not showing when 'disabled'
The locale selector was not being displayed when locale is set to
'disabled', and useAppsGallery is true in the settings.yml. The reason
for that is because when the useAppsGallery is set, the transcriptions
are enabled or disabled via toggle and not by selecting the disabled
option in a selector dropdown. So to fix that, a parameter was added to
the getLocale function, which allows skipping the locale if it is set to
'disabled' and returning the most similar available locale to the
navigator language.