feat(livekit): add room connection retry
Add a manual LK room connection retry on startup. The retry will kick
whenever a connection error is fired by the room's onError callback, as
well as meet all the criteria for the reconn to be valid (e.g.: client
is connected, conn info is present, etc). Max retry attempts is 10.