fix(audio): extend retryable gUM error list
A portion of audio init errors are NotFoundError-coded gUM failures.
General cause for that is having no input devices, but sometimes they
_are_ present and this error is thrown anyways. The hypothesis here is
that tainted constraints (which carry deviceId and other attributes from
previous sessions) might be causing this. Same goes for NotReadableError.

Add NotFoundError and NotReadableError to the list of retryable gUM
errors for audio. I'll keep an eye on telemetry data to see if this has
any effect.