fix(audio): render loop in audio settings if webspeech is unavailable
The audio settings modal causes a render loop in browsers that don't
fully support the WebSpeech API.

This originates from the underlying AudioCaptionsSelect component. When
migrated to TS, an issue was introduced where an the voicesList state
is reset on a loop due to it not being in the effect's dependency array
nor checking if the WebSpeech voices array is empty in the first place.

Prevent the render loop by:
  - Checking if the voices array is empty before setting it
  - Adding the voicesList to the effect's dependecy array so it doesn't
    use stale data