fix(screenshare): remove instability UI
Screen sharing has a dedicated instability UI that is rendered
whenever the peer state transitions to an unhealthy state and packets
stop flowing for an specified interval. We've heard multiple feedbacks
from both admins and end users alike that this is too sensitive and
disruptive to be of any help. Additionally:
  - It's redundant as the conn status component already pulls similar
    data (including other RTC sources) and reacts to it accordingly
  - It's inconsistent as neither cameras nor audio have such dedicated
    UIs (both rely on conn status as well)
  - ICE restart being implemented means reconns happen faster - so the
    UI is less useful

Remove screen share's instability UI completely which should decrease
the number of unwarranted disruptions during screen sharing sessions.
Further improvements to reconn detection and notification should be done
later with the assistance of the design team, preferably.